First version a little stable (alpha): - Fixed makefile installation not own - Simplification of Man-in-the-middle mode usage and dump - Implementation of all sub layer UDP process, to resend/ack packets and handle possible errors exchange - Implementation of NetGroup.close message - Handle ping peer value - New comer peer for a NetGroup received now the best peer as its entry point (with the lowest ping value) - Discovery of peer death accelerated and optimized - Clean correctly the RTMFPServer instance after a stop - MemoryStream written bug fixed - Clean empty NetGroup during the server process - Optimization of session death - Keepalive server implemented - On stop server, the clients are now advised - Fixed some other normal mode bugs - Fixed some Man-in-the-middle mode bugs - Added of TODO file - Readme updated

+++ CumulusLib/include/ServerData.h
@@ -25,7 +25,7 @@ namespace Cumulus {
 class CUMULUS_API ServerData
 {
 public:
-	ServerData(Poco::UInt16 keepAliveServer,Poco::UInt16 keepAlivePeer);
+	ServerData(Poco::UInt8 keepAliveServer,Poco::UInt8 keepAlivePeer);
 	virtual ~ServerData();
 
 	Group& group(const std::vector<Poco::UInt8>& id);
